+++ composeApp/src/commonMain/kotlin/com/vivimusic/de/data/AppContainer.kt
@@ -4,6 +4,8 @@ import com.vivimusic.de.data.db.AppDatabase
 import com.vivimusic.de.data.db.getRoomDatabase
 import com.vivimusic.de.data.network.InnerTubeClient
 import com.vivimusic.de.data.network.createHttpClient
+import com.vivimusic.de.data.playback.AudioEngine
+import com.vivimusic.de.data.playback.createAudioEngine
 import com.vivimusic.de.data.sync.SupabaseSyncClient
 import com.vivimusic.de.data.sync.SyncManager
 import kotlinx.coroutines.CoroutineScope
@@ -26,6 +28,8 @@ class AppContainer(val scope: CoroutineScope) {
     val syncManager: SyncManager = SyncManager(database, syncClient, scope)
 
     val repository: MusicRepository = MusicRepository(database, innerTube, syncManager, scope)
+
+    val audioEngine: AudioEngine = createAudioEngine()
 
     fun start() {
         if (syncManager.isEnabled) {